refactor(audio): extract shared audio primitives into src/audio layer
WavEncoder, AudioEncoder, AudioFormatConverter, streamingConversion,
EncodingWorkerClient, encodingWorker, AudioCapabilityDetector, and
RecordingFileManager served as codec/format/file primitives for every
feature, but lived inside the recording feature, creating cross-feature
edges (cleanup -> recording, transcription -> recording, ui/settings/
diagnostics -> recording internals).

They now form a dedicated src/audio layer that recording, cleanup,
transcription, and the UI all depend on. Pure path move plus import
rewrites; no behavior change. The esbuild worker entry point follows
the encodingWorker move.
