+++ src/recording/PartRotationController.ts
@@ -14,7 +14,7 @@ import type { RecordingSessionConfig, RecordingTarget } from '../types';
 import type { AudioRecorderSettings } from '../settings/Settings';
 import { PLUGIN_LOG_PREFIX, MS_PER_MINUTE, FORMAT_WAV } from '../constants';
 import { DebugLogger } from '../utils/DebugLogger';
-import { resolveUniquePath } from './RecordingFileManager';
+import { resolveUniquePath } from '../audio/RecordingFileManager';
 import {
 	buildPartFileName,
 	detachTrailingBytes,
