+++ tests/unit/ConversionService.test.ts
@@ -23,13 +23,13 @@ jest.mock('obsidian', () => ({
 	normalizePath: (path: string) => path.replace(/\\/g, '/'),
 }));
 
-jest.mock('../../src/recording/AudioEncoder', () => ({
+jest.mock('../../src/audio/AudioEncoder', () => ({
 	encodeAudioBuffer: jest
 		.fn()
 		.mockResolvedValue(new Blob(['encoded'], { type: 'audio/wav' })),
 }));
 
-jest.mock('../../src/recording/AudioFormatConverter', () => ({
+jest.mock('../../src/audio/AudioFormatConverter', () => ({
 	decodeAudioBlob: jest.fn().mockResolvedValue({}),
 	convertBlobToFormat: jest
 		.fn()
@@ -122,10 +122,10 @@ describe('ConversionService', () => {
 
 	it('should use the decode-and-encode path for WAV targets', async () => {
 		const { decodeAudioBlob } = jest.requireMock(
-			'../../src/recording/AudioFormatConverter',
+			'../../src/audio/AudioFormatConverter',
 		);
 		const { encodeAudioBuffer } = jest.requireMock(
-			'../../src/recording/AudioEncoder',
+			'../../src/audio/AudioEncoder',
 		);
 
 		await service.convert(
