+++ tests/unit/RecordingFinalizer.test.ts
@@ -21,19 +21,19 @@ jest.mock('obsidian', () => ({
 	normalizePath: (path: string) => path.replace(/\\/g, '/'),
 }));
 
-jest.mock('../../src/recording/WavEncoder', () => ({
+jest.mock('../../src/audio/WavEncoder', () => ({
 	assembleWavFromPcmSegmentFiles: jest
 		.fn()
 		.mockResolvedValue(new ArrayBuffer(50)),
 }));
 
-jest.mock('../../src/recording/AudioEncoder', () => ({
+jest.mock('../../src/audio/AudioEncoder', () => ({
 	isOfflineEncodingSupported: jest.fn((format: string) =>
 		['mp3', 'flac', 'webm', 'ogg', 'mp4', 'm4a', 'aac'].includes(format),
 	),
 }));
 
-jest.mock('../../src/recording/AudioFormatConverter', () => ({
+jest.mock('../../src/audio/AudioFormatConverter', () => ({
 	mergeAudioTracks: jest
 		.fn()
 		.mockResolvedValue(new Blob(['merged'], { type: 'audio/wav' })),
@@ -219,7 +219,7 @@ describe('RecordingFinalizer', () => {
 		it('should convert to WAV when the output format is wav', async () => {
 			buildFinalizer(createSession({ outputFormat: 'wav' }));
 			const { convertBlobToWav } = jest.requireMock(
-				'../../src/recording/AudioFormatConverter',
+				'../../src/audio/AudioFormatConverter',
 			);
 
 			await finalizer.finalizeSegmentsToFile(['seg1.tmp'], 'final.wav');
@@ -230,7 +230,7 @@ describe('RecordingFinalizer', () => {
 		it('should re-encode offline-only formats with remux allowed and mapped progress', async () => {
 			buildFinalizer(createSession({ outputFormat: 'mp3' }));
 			const { convertBlobToFormat } = jest.requireMock(
-				'../../src/recording/AudioFormatConverter',
+				'../../src/audio/AudioFormatConverter',
 			);
 
 			await finalizer.finalizeSegmentsToFile(
@@ -280,7 +280,7 @@ describe('RecordingFinalizer', () => {
 	describe('assembleWavFile', () => {
 		it('should delegate assembly to the shared single-allocation helper', async () => {
 			const { assembleWavFromPcmSegmentFiles } = jest.requireMock(
-				'../../src/recording/WavEncoder',
+				'../../src/audio/WavEncoder',
 			);
 			const target = createTarget({
 				segmentPaths: ['pcm1.tmp', 'pcm2.tmp'],
@@ -524,7 +524,7 @@ describe('RecordingFinalizer', () => {
 			await finalizer.saveRecording(targets, 'stamp', null);
 
 			const { mergeAudioTracks } = jest.requireMock(
-				'../../src/recording/AudioFormatConverter',
+				'../../src/audio/AudioFormatConverter',
 			);
 			expect(mergeAudioTracks).not.toHaveBeenCalled();
 			const { insertFileLinks } = jest.requireMock(
@@ -567,7 +567,7 @@ describe('RecordingFinalizer', () => {
 				'../../src/recording/StreamingMixer',
 			);
 			const { mergeAudioTracks } = jest.requireMock(
-				'../../src/recording/AudioFormatConverter',
+				'../../src/audio/AudioFormatConverter',
 			);
 			buildFinalizer(
 				createSession({
@@ -616,7 +616,7 @@ describe('RecordingFinalizer', () => {
 				'../../src/recording/StreamingMixer',
 			);
 			const { mergeAudioTracks } = jest.requireMock(
-				'../../src/recording/AudioFormatConverter',
+				'../../src/audio/AudioFormatConverter',
 			);
 			(canStreamMix as jest.Mock).mockReturnValueOnce(false);
 			buildFinalizer(
@@ -652,7 +652,7 @@ describe('RecordingFinalizer', () => {
 				'../../src/recording/StreamingMixer',
 			);
 			const { mergeAudioTracks } = jest.requireMock(
-				'../../src/recording/AudioFormatConverter',
+				'../../src/audio/AudioFormatConverter',
 			);
 			buildFinalizer(
 				createSession({
